requirements.txt
PyQt6
Pillow
piexif
imagehash
imagehash
numpy
opencv-python
